Prepare multiple environments
With multiple different CPUs multiple environment and testing
programs need to be created. This commit prepares the codebase to
work with different environments set in conf.lua.

Environment
Each environment is a directory containing a main.lua as
entrypoint to the environment and a conf.lua with a `loveconf(t)`
function that extends the base love.conf function with environment
specific settings.