Fix error message on device adding
When a new device is added to the BUS, the BUS verifies that there
is enough space on the address range to put the device in. If that
fails, an error is thrown.

The first fix of this commit is that string.format didn't receive
one of the arguments it requires to print out the error message.

The second fix is how the BUS handled the CPU declaring its
available address range. 0 was not being counted as a valid
address which removed the visible CPU address space to the
BUS by 1.