Fix relative addressing: Part 2
A mistake that was made during the first fix of the REL addressing
mode is that Lua considers `0` a truthy value, compared to what
many other languages do, aka `0` being falsy. This caused even
positive values to be shifted and lead to wrong branching. By
specifically checking non-equality to `0`, this problem is fixed.