Fix wrong JMP address in example program
The tests have been re-aligned but the JMP instruction's address
was still set to the old address, causing a nonsensical jump.
The address has been fixed to point to the right page.

Moreover, the disassembler has been changed to only show
instructions after the current instruction. This eliminates wrong
guesses of instructions that come before the current instruction.